requirements.txt
python-telegram-bot==21.0
httpx==0.27.0
pyrogram==2.0.106
aiosqlite==0.20.0
loguru==0.7.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
TgCrypto==1.2.5
